Do not compute hash via X-Forwarded-Proto
This line fixed Varnish caching HTTPS ugrades, however by factoring in this header when computing the hash we've caused PURGE requests to be ineffective: The PURGE matches the HTTP requests, but not HTTPS requests. This isn't useful.

The correct way to handle this is to have the back-end (which is doing the HTTPS ugprade) inform Varnish to vary the cache upon X-Forwarded-For

fixes #2